Change SPEC4 to cover using and building nightly wheels (#195)
The goal is to create a SPEC that covers both using and creating nightlies.

I am trying to keep it kind of high-level, giving examples of how to do the things that are new (e.g. uploading to the nightly area) but not be too prescriptive/show full blown CI setups because I assume the target audience are likely to already have fairly sophisticated/complex/complicated CI setups. For example scikit-learn uses Azure to run their unit tests, but uses GH Actions to build wheels. I think this is good, we don't want people to revamp their whole CI setup.