Change to the JSON version of the PyPi Simple API
According to the PyPa API Reference, the `releases` key of the JSON
API is deprecated:

https://warehouse.pypa.io/api-reference/json.html

Confusingly, the Simple API is recommended. The Simple API was
originally described in PEP 503 (2015), but an updated version was proposed
in PEP 691 (2022) that adds a JSON response.

However, this API does *not* include the same information as in the
`releases` key of the current JSON API. But the PyPa authors are
unlikely to add it back:

https://github.com/pypi/warehouse/issues/11991

So, this PR instead follows their recommended path, which is to parse versions
from uploaded filenames (!).

Closes #218