Re-wording of admin selection recommendations.
A minor rewording proposal to clarify the underlying motivations
of the admin selection process. I also tried to soften the wording
so that it wouldn't seem as prescriptive so it's clear that the
decision-making processes amongst the existing admins is decoupled
from the process used by individual projects to nominate
representatives.