Generate gallery from GH teams (#276)
The spec steering committee section needs to be updated to explain how
to add members, but that will be done in a separate PR as I think we
need to have a general section about using the teams repo to add members
and teams. Then I will update this to point to that documentation.

But just to capture it somewhere for now, the basic process is:
- make a PR to https://github.com/scientific-python/teams
- make a PR to
https://github.com/scientific-python/scientific-python.org/ after
running `make teams`
- add them to
https://discuss.scientific-python.org/g/spec-steering-committee

It requires that they have accounts on both GitHub and
discuss.scientific-python.org.