Add publish state to SPEC process (#339)
Motivation: while writing SPECs, and shopping them around, I've noticed
some confusion around when SPECs are ready for endorsement. Currently,
we have some DRAFT specs that are ready for endorsements, and others
that have not been developed at all.

The clarification made by this PR is to introduce an editing phase, so
that all SPECs start out as drafts. They are not ready for endorsement
in this phase. Authors edit their drafts, and once they are ready to
seek endorsement, they **publish** their SPEC (on the website, the DRAFT
flag & warning disappears). Now, they can canvass for endorsement.