Add SPEC9 - Governance (#323)
Hi everyone! 👋🏻 

This PR adds the first draft for SEPC9 — Governance developed during the
[Scientific Python Developer Summit
2024](https://scientific-python.org/summits/developer/2024/).

This SPEC outlines the definition of governance, various open-source
governance models, the steps to choose the right governance model, and
what needs to be done post-adoption.

I'd appreciate feedback and thoughts on the current draft. Thanks!

---------

Co-authored-by: Paul Ivanov <pi@berkeley.edu>
Co-authored-by: Inessa Pawson <inessapawson@gmail.com>
Co-authored-by: Daniel McCloy <dan@mccloy.info>
Co-authored-by: Matt Haberland <mhaberla@calpoly.edu>
Co-authored-by: Brigitta Sipőcz <bsipocz@gmail.com>
Co-authored-by: Brigitta Sipőcz <b.sipocz@gmail.com>